Former-commit-id: a65ed8460107c7d27d340290df64e6030d7614aa

+++ back/src/main/java/bunsan/returnz/persist/entity/GamerStock.java
@@ -41,7 +41,7 @@ public class GamerStock {
 	@Builder.Default
 	@ColumnDefault("0")
 	@Column(nullable = false)
-	private Double averagePrice = 0.0;
+	private Integer averagePrice = 0;
 
 	@Builder.Default
 	@ColumnDefault("0")
